Create GenerateFile2023
This script will generate seven lines of random data, each labeled with a number. Then, it will delete five random lines from the generated data. The final result will be saved in a file named data.txt. You can modify the script to adjust the number of lines to generate and delete as per your requirements.